refactor: 수익율 계산하는 로직을 lotteryStatisticsGenerator으로 이동
