+++ src/main/kotlin/lotto/business/LotteryStatisticsGenerator.kt
@@ -8,4 +8,9 @@ class LotteryStatisticsGenerator {
         val sixMatchCount = prizeResults.prizeCountMap[LotteryPrize.SIX_MATCH] ?: 0
         return LotteryStatistics(treeMatchCount, fourMatchCount, fiveMatchCount, sixMatchCount)
     }
+
+    fun calculateProfitRate(receivedAmount: ReceivedAmount, prizeResults: PrizeResults): ProfitRate {
+        val totalPrize = prizeResults.prizeCountMap.map { it.key.prizeAmount * it.value }.sum()
+        return ProfitRate(totalPrize / receivedAmount.amount.toDouble())
+    }
 }
