+++ src/main/kotlin/lotto/business/PrizeResults.kt
@@ -1,8 +1,3 @@
 package lotto.business
 
-class PrizeResults(val prizeCountMap: Map<LotteryPrize, Int>) {
-    fun calculateProfitRate(receivedAmount: ReceivedAmount): ProfitRate {
-        val totalPrize = prizeCountMap.map { it.key.prizeAmount * it.value }.sum()
-        return ProfitRate(totalPrize / receivedAmount.amount.toDouble())
-    }
-}
+class PrizeResults(val prizeCountMap: Map<LotteryPrize, Int>)
