+++ src/test/kotlin/lotto/business/LotteryStatisticsGeneratorTest.kt
@@ -65,4 +65,26 @@ class LotteryStatisticsGeneratorTest {
             LottoNumber(num6)
         )
     )
+
+    @Test
+    fun `받은 금액과 비교하여 수익률을 계산한다`() {
+        // given
+        val prizeResults = PrizeResults(
+            mapOf(
+                LotteryPrize.THREE_MATCH to 5,
+                LotteryPrize.FOUR_MATCH to 4,
+                LotteryPrize.FIVE_MATCH to 3,
+                LotteryPrize.SIX_MATCH to 1
+            )
+        )
+        val receivedAmount = ReceivedAmount(100000000)
+        val expected = ProfitRate(20.04725)
+        val lotteryStatisticsGenerator = LotteryStatisticsGenerator()
+
+        // when
+        val profitRate = lotteryStatisticsGenerator.calculateProfitRate(receivedAmount, prizeResults)
+
+        // then
+        assertThat(profitRate).isEqualTo(expected)
+    }
 }
